requirements.txt
streamlit==1.37.1
pandas==2.2.2
scikit-learn==1.5.1
python-dotenv==1.0.1
nltk==3.8.1
requests==2.32.3
